requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
PyMuPDF>=1.23.0
pymongo>=4.6.0
mcp>=1.0.0
anthropic>=0.40.0
openai>=2.0.0
google-genai>=1.0.0
